Move functions in `airflow.utils.decorator` to more appropriate places
`classproperty` was only used in airflow.models.deadline.py so I moved it
there. We can find a better place for it when there are more use of it outside
the deadline module
