+++ airflow-core/src/airflow/utils/__init__.py
@@ -38,6 +38,10 @@ __deprecated_classes = {
         # should be in user code or custom providers, so redirect them to the public interface in Task SDK
         "*": "airflow.sdk.timezone"
     },
+    "decorators": {
+        "remove_task_decorator": "airflow.sdk.definitions._internal.decorators.remove_task_decorator",
+        "fixup_decorator_warning_stack": "airflow.sdk._internal.decorators.fixup_decorator_warning_stack",
+    },
 }
 
 add_deprecated_classes(__deprecated_classes, __name__)
