+++ providers/standard/tests/unit/standard/utils/test_python_virtualenv.py
@@ -24,7 +24,7 @@ from unittest import mock
 import pytest
 
 from airflow.providers.standard.utils.python_virtualenv import _generate_pip_conf, _use_uv, prepare_virtualenv
-from airflow.utils.decorators import remove_task_decorator
+from airflow.sdk.bases.decorator import remove_task_decorator
 
 from tests_common.test_utils.config import conf_vars
